+++ frontend/frontend.php
@@ -96,3 +96,81 @@ function cpt_no_title_title_tag( $title_parts ) {
 
 	return $title_parts;
 }
+
+if ( get_option( 'cpt_sites_open_external_links_in_new_tab' ) ) {
+	add_filter( 'the_content', 'cpt_open_external_links_in_new_tab' );
+	add_filter( 'the_excerpt', 'cpt_open_external_links_in_new_tab' );
+	add_filter( 'widget_text', 'cpt_open_external_links_in_new_tab' );
+	add_filter( 'widget_custom_html', 'cpt_open_external_links_in_new_tab' );
+	add_filter( 'wp_nav_menu', 'cpt_open_external_links_in_new_tab' );
+	add_filter( 'comment_text', 'cpt_open_external_links_in_new_tab' );
+}
+/**
+ * Adds `target="_blank"` to external links in the supplied HTML.
+ *
+ * @param string $content An HTML string.
+ * @return string
+ */
+function cpt_open_external_links_in_new_tab( $content ) {
+	// Exit if there are no links.
+	if ( stripos( $content, '<a ' ) === false ) {
+		return $content;
+	}
+
+	$this_website = wp_parse_url( home_url(), PHP_URL_HOST );
+	$show_icon    = get_option( 'cpt_sites_show_external_link_icon' );
+
+	// Suppress HTML parsing errors.
+	libxml_use_internal_errors( true );
+
+	$doc = new DOMDocument();
+	// Load $doc as a UTF-8 HTML fragment.
+	$doc->loadHTML(
+		'<?xml encoding="UTF-8">' . $content,
+		LIBXML_HTML_NOIMPLIED | LIBXML_HTML_NODEFDTD
+	);
+
+	$links = $doc->getElementsByTagName( 'a' );
+
+	foreach ( $links as $link ) {
+		$href = $link->getAttribute( 'href' );
+		if ( empty( $href ) || strpos( $href, 'http' ) !== 0 ) {
+			continue; // Skip non-http links (anchors, mailto, etc.).
+		}
+
+		$link_website = wp_parse_url( $href, PHP_URL_HOST );
+
+		// Add `target="_blank"` if the link is external.
+		if ( $link_website && $link_website !== $this_website && strpos( $link_website, $this_website ) === false ) {
+			$link->setAttribute( 'target', '_blank' );
+
+			// Add or merge `rel="noopener, noreferrer"` attributes.
+			$existing_rel  = $link->getAttribute( 'rel' );
+			$required_rels = array( 'noopener', 'noreferrer' );
+
+			$rels = array_unique(
+				array_filter(
+					array_merge(
+						preg_split( '/\s+/', $existing_rel, -1, PREG_SPLIT_NO_EMPTY ),
+						$required_rels
+					)
+				)
+			);
+
+			$link->setAttribute( 'rel', implode( ' ', $rels ) );
+
+			// Add external link icon if option is enabled.
+			if ( $show_icon ) {
+				$icon = $doc->createElement( 'i' );
+				$icon->setAttribute( 'class', 'cpt-icon external-link' );
+				$link->appendChild( $icon );
+			}
+		}
+	}
+
+	$html = $doc->saveHTML();
+
+	libxml_clear_errors();
+
+	return $html;
+}